chore: cleanup DataTableColumnHeader (#4540)
Some refactoring to `DataTableColumnHeader` to make it more easily
extendible without merge-conflicts
