+++ frontend/src/css/common.css
@@ -32,3 +32,7 @@
 :root:has(:fullscreen) .hide-on-fullscreen {
   display: none !important;
 }
+
+.mo-dropdown-icon {
+  @apply mr-2 h-3.5 w-3.5 text-muted-foreground/70;
+}
